bug: Body undefined
Uses the document.body, fully sopported on modern browsers in
replacement of a simple query on the DOM. In some SPAs, like Vue, the
body DOM element does not exist when making the DOM query. The
document.body instead, it exists and can be used to attach the script
DOM element.

See: KJS-965